+++ docs-src/src/pages/index.tsx
@@ -1,7 +1,5 @@
-import { Link } from 'gatsby';
-import React, { useState } from 'react';
+import React from 'react';
 
-import { DrawerComponent } from '../components/Drawer';
 import Layout from '../components/layout';
 import SEO from '../components/seo';
 
@@ -10,8 +8,6 @@ const IndexPage = () => {
   return (
     <Layout>
       <SEO title="Home" keywords={[`infoSum`, `charts`, `react`]} description="" />
-
-
     </Layout>
   );
 };